+++ examples/solid/basic-virtual-file-based/src/routeTree.gen.ts
@@ -178,7 +178,7 @@ declare module '@tanstack/solid-router' {
     '/_first': {
       id: '/_first'
       path: ''
-      fullPath: ''
+      fullPath: '/'
       preLoaderRoute: typeof layoutFirstLayoutRouteImport
       parentRoute: typeof rootRouteImport
     }
@@ -199,7 +199,7 @@ declare module '@tanstack/solid-router' {
     '/_first/_second-layout': {
       id: '/_first/_second-layout'
       path: ''
-      fullPath: ''
+      fullPath: '/'
       preLoaderRoute: typeof layoutSecondLayoutRouteImport
       parentRoute: typeof layoutFirstLayoutRoute
     }