+++ packages/router-cli/src/index.ts
@@ -1,4 +1,4 @@
-import * as yargs from 'yargs'
+import yargs from 'yargs'
 import { getConfig } from '@tanstack/router-generator'
 import { generate } from './generate'
 import { watch } from './watch'
@@ -6,8 +6,7 @@ import { watch } from './watch'
 main()
 
 export function main() {
-  yargs
-    // @ts-expect-error
+  yargs()
     .scriptName('tsr')
     .usage('$0 <cmd> [args]')
     .command('generate', 'Generate the routes for a project', async () => {