+++ packages/router-devtools-core/vite.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig, mergeConfig } from 'vitest/config'
-import { tanstackViteConfig } from '@tanstack/config/vite'
+import { tanstackViteConfig } from '@tanstack/vite-config'
 import solid from 'vite-plugin-solid'
 
 const config = defineConfig({
@@ -15,7 +15,7 @@ const merged = mergeConfig(
   }),
 )
 
-merged.build.rollupOptions.output.manualChunks = undefined
-merged.build.rollupOptions.output.preserveModules = false
+merged.build.rolldownOptions.output.manualChunks = undefined
+merged.build.rolldownOptions.output.preserveModules = false
 
 export default merged