+++ packages/start-static-server-functions/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @tanstack/start-static-server-functions
 
+## 1.166.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.166.11
+  - @tanstack/react-start@1.166.14
+  - @tanstack/solid-start@1.166.14
+
+## 1.166.12
+
+### Patch Changes
+
+- build: update to vite-config 5.x (rolldown) ([#6926](https://github.com/TanStack/router/pull/6926))
+
+- Updated dependencies [[`838b0eb`](https://github.com/TanStack/router/commit/838b0eb9a8bbbb987a0a6972c1446e01423bbd7b)]:
+  - @tanstack/react-start@1.166.13
+  - @tanstack/solid-start@1.166.13
+  - @tanstack/start-client-core@1.166.10
+
+## 1.166.11
+
+### Patch Changes
+
+- fix: build with @tanstack/vite-config 0.4.3 ([#6923](https://github.com/TanStack/router/pull/6923))
+
+- Updated dependencies [[`ef9b241`](https://github.com/TanStack/router/commit/ef9b241f3cfe95cee40daa96da669f0ffd4a971a)]:
+  - @tanstack/start-client-core@1.166.9
+  - @tanstack/react-start@1.166.12
+  - @tanstack/solid-start@1.166.12
+
+## 1.166.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/start-client-core@1.166.8
+  - @tanstack/react-start@1.166.10
+  - @tanstack/solid-start@1.166.10
+
 ## 1.166.8
 
 ### Patch Changes